changing to use Router
