+++ First-React-App/src/components/AddCount/AddCount.css
@@ -1,0 +1,7 @@
+.count-button {
+    background-color: rgb(90, 121, 215);
+}
+
+.count-button:hover {
+    background-color: rgb(70, 101, 195);
+}
