+++ First-React-App/src/pages/HelpPage.jsx
@@ -1,0 +1,15 @@
+import { Router, useNavigate } from "react-router-dom"
+
+export default function HelpPage() {
+    const navigate = useNavigate();
+
+    const goToHome = () => {
+        navigate("/");
+    }
+    return (
+        <>
+            <h3>BAM</h3>
+            <button onClick={() => goToHome("/")}>Click</button>
+        </>
+    );
+}
